Split Parser module which have 2 or more  parser in one module.
should 1 Parser 1 module now, submodule is counted as module.
