+++ src/main/java/jcomicdownloader/module/Parse99770.java
@@ -1,0 +1,23 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package jcomicdownloader.module;
+
+import jcomicdownloader.enums.Site;
+
+/**
+ *
+ * @author apple
+ */
+public class Parse99770 extends ParseNINENINE
+{
+    public Parse99770()
+    {
+        super();
+        enumName = "NINENINE_99770";
+        parserName=this.getClass().getName();
+        siteID=Site.formString("NINENINE_99770");
+    }
+}
